// TXC-MicroFrontends/Voucher/src/app/voucher/enum/inventory-sku-details-action-enum.ts
export enum InventorySkuDetailsActionEnum {
    DOWNLOAD_INVENTORY = 'Download inventory',
    ADD_STOCK_QTY = 'Add stock quantity',
    TRASH_INVENTORY = 'Trash inventory',
}

export enum InventoryBatchDetailsActionEnum {
    TRASH_BATCH = 'Trash batch',
    DELETE_BATCH = 'Delete batch',
    DOWNLOAD_BATCH = 'Download batch',
    EDIT_DATE_INFO = 'Edit date info',
}